Fixes for fitting and examples
