+++ easyCore/Fitting/bumps.py
@@ -72,7 +72,7 @@ class bumps(FittingTemplate):  # noqa: S101
                         par[
                             "p" + str(NameConverter().get_key(item))
                         ] = obj.convert_to_par_object(item)
-                return Curve(fit_func, x, y, weights, **par)
+                return Curve(fit_func, x, y, dy=weights, **par)
 
             return make_func
 
@@ -180,16 +180,16 @@ class bumps(FittingTemplate):  # noqa: S101
         if method is not None and method in self.available_methods():
             default_method["method"] = method
 
-        if weights is None:
-            weights = np.sqrt(np.abs(y))
+        # if weights is None:
+        #     weights = np.sqrt(np.abs(y))
 
         if engine_kwargs is None:
             engine_kwargs = {}
 
         if minimizer_kwargs is None:
             minimizer_kwargs = {}
-        else:
-            minimizer_kwargs = {"fit_kws": minimizer_kwargs}
+        # else:
+        #     minimizer_kwargs = {"fit_kws": minimizer_kwargs}
         minimizer_kwargs.update(engine_kwargs)
 
         if model is None:
@@ -206,7 +206,7 @@ class bumps(FittingTemplate):  # noqa: S101
 
         borg.stack.beginMacro("Fitting routine")
         try:
-            model_results = bumps_fit(problem, **kwargs)
+            model_results = bumps_fit(problem, **default_method, **minimizer_kwargs, **kwargs)
             self._set_parameter_fit_result(model_results)
             results = self._gen_fit_results(model_results)
         except Exception as e:
@@ -291,6 +291,7 @@ class bumps(FittingTemplate):  # noqa: S101
 
         results.fitting_engine = self.__class__
         results.fit_args = None
+        results.engine_result = fit_results
         # results.check_sanity()
         return results
