+++ examples_old/example3_broken.py
