+++ examples_old/example_dataset4_2.py
@@ -10,36 +10,44 @@ import matplotlib.pyplot as plt
 
 d = xr.Dataset()
 
+
 class Wavey(BaseObj):
     def __init__(self,
                  amplitude: Parameter,
                  phase: Parameter,
                  period: Parameter):
-        super(Wavey, self).__init__(amplitude=amplitude, phase=phase, period=period)
+        super(Wavey, self).__init__('SimpleWave',
+                                    amplitude=amplitude, phase=phase,
+                                    period=period)
+
+    def __repr__(self):
+        return (f'{self.name} - A={self.amplitude}, ph={self.phase}, period={self.period}')
 
     @classmethod
     def from_params(cls, amplitude: float = 1, phase: float = 0, period: float = 2*np.pi):
         amplitude = Parameter('amplitude', amplitude, min=0)
         phase = Parameter('phase', phase)
-        period = Parameter('period', period)
+        period = Parameter('period', period, min=0, max=2*np.pi)
         return cls(amplitude, phase, period)
 
     def fit_fun(self, x, *args, **kwargs):
         # In the real case we would gust call the evaluation fn without reference to the BaseObj
         return self.amplitude.raw_value * np.sin((x + self.phase.raw_value)/self.period.raw_value)
 
+b = Wavey.from_params()
+bb = Wavey.from_params(1.1, 0.1, 1.9*np.pi)
 f = Fitter()
 
-f.initialize(b, fit_fun)
+f.initialize(b, b.fit_fun)
 
 nx = 1E3
 x_min = 0
 x_max = 100
 
 x = np.linspace(x_min, x_max, num=int(nx))
-y1 = 2*x - 1 + 5*(np.random.random(size=x.shape) - 0.5)
+y1 = bb.fit_fun(x) + .75*(0.5 - np.random.random(size=x.shape))
 x2 = x + 20
-y2 = 2*x2 - 1 + 5*(np.random.random(size=x2.shape) - 0.5)
+y2 = bb.fit_fun(x2) + .75*(0.5 - np.random.random(size=x2.shape))
 
 d.easyCore.add_coordinate('x1', x)
 d.easyCore.add_variable('y1', ['x1'], y1, auto_sigma=True)
@@ -54,4 +62,5 @@ for idx, r in enumerate(res):
     r.y_calc.plot(ax=axs[idx])
     axs[idx].set_title(f'Dataset {idx}')
 plt.show()
-
+print(f'Actual = {bb}')
+print(f'Fitted = {b}')
